+++ src/main/java/act/inject/param/PartialPathLoader.java
@@ -41,4 +41,8 @@ public class PartialPathLoader implements ParamValueLoader {
         return bindName;
     }
 
+    @Override
+    public boolean supportJsonDecorator() {
+        return false;
+    }
 }