+++ src/main/java/act/inject/param/SessionValueLoader.java
@@ -68,6 +68,11 @@ class SessionValueLoader implements ParamValueLoader {
         return key;
     }
 
+    @Override
+    public boolean supportJsonDecorator() {
+        return false;
+    }
+
     private String key(String name, BeanSpec spec) {
         if (S.notBlank(name)) {
             return name;